Add Support for Config Inheritance
Adds support for a top-level `parent` key in configuration files which
will merge the contents of that path with the contents of the current
file. The merge rules are the same as specifying multiple configuration
files at the command line with the "parent" config being loaded first.

This also moves configuration loading out into the `splat.conf` module
so configuration loading can be used as a library by other projects.